Flatten dock Export panel — remove card wrapper to fit in dock
The export_panel used card_with_header which added frame padding
inside the already-padded dock, causing content to overflow even
with increased min_height. Inline a flat version in dock_export
and remove the now-unused export_panel wrapper.
